[webOS5.0][K6LP][8812] Checking Coverity issue
:Release Notes:
[webOS5.0][K6LP][8812] Checking Coverity issue

:Detailed Notes:
High Impact/ Medium Impac
4 (kernel) / 1 (kernel)

:Testing Performed:
check on K6LP

:Issues Addressed:
(RTKWBTV-730)
	[webOS5.0][K6LP][8812] Checking Coverity issue

+++ hal/phydm/phydm_adaptivity.h
@@ -101,7 +101,7 @@ struct phydm_adaptivity_struct {
 	u16			igi_up_bound_lmt_val;	/*max value of igi_up_bound_lmt_cnt*/
 	boolean		igi_lmt_en;
 	u8			adapt_igi_up;
-	s8			rvrt_val[2];
+	u32			rvrt_val[2];
 	s8			th_l2h;
 	s8			th_h2l;
 };
